Split debtCache tests
