+++ contracts/DebtCache.sol
@@ -2,11 +2,10 @@ pragma solidity ^0.5.16;
 
 // Inheritance
 import "./BaseDebtCache.sol";
-import "./interfaces/IDebtCache.sol";
 
 
 // https://docs.synthetix.io/contracts/source/contracts/debtcache
-contract DebtCache is BaseDebtCache, IDebtCache {
+contract DebtCache is BaseDebtCache {
     constructor(address _owner, address _resolver) public BaseDebtCache(_owner, _resolver) {}
 
     /* ========== MUTATIVE FUNCTIONS ========== */
