+++ contracts/interfaces/IBaseDebtCache.sol
@@ -35,4 +35,18 @@ interface IBaseDebtCache {
             bool isInvalid,
             bool isStale
         );
+
+    // Mutative functions
+
+    function takeDebtSnapshot() external;
+
+    function updateCachedSynthDebts(bytes32[] calldata currencyKeys) external;
+
+    function purgeCachedSynthDebt(bytes32 currencyKey) external;
+
+    function updateCachedSynthDebtWithRate(bytes32 currencyKey, uint currencyRate) external;
+
+    function updateCachedSynthDebtsWithRates(bytes32[] calldata currencyKeys, uint[] calldata currencyRates) external;
+
+    function updateDebtCacheValidity(bool currentlyInvalid) external;
 }
